* All tools:     - Change: adopt a tree layout compliant with the UNIX FHS.         * includes under      TOP/include/coriolis2.         * shared datas under  TOP/shared/coriolis2.         * docs under          TOP/share/doc/coriolis2.         * configuration under TOP/etc/coriolis2
  * ./crlcore:
    - Change: In Environment, comply to the new tree layout, search configuration
        files under TOP/etc/coriolis2/.

  * ./knik:
    - Change: In flute, comply to the new tree layout, get the "POW*.dat" files
        from TOP/share/coriolis2/flute-2.4.

+++ hurricane/src/viewer/RecordModel.cpp
@@ -126,9 +126,9 @@ namespace Hurricane {
     }
 
     if ( role == Qt::DisplayRole ) {
-      int   row  = index.row ();
+      int row = index.row ();
     //Slot* slot = _record->getSlot ( row );
-      if ( row < _cache.size() ) {
+      if ( row < (int)_cache.size() ) {
         switch ( index.column() ) {
           case 0: return _cache[row].first;
           case 1: return _cache[row].second;
